+++ examples/json.py
@@ -1,12 +1,9 @@
-import re
 from sys import stdin
 
 from parsy import generate, regex, string
 
-whitespace = regex(r'\s*', re.MULTILINE)
-
+whitespace = regex(r'\s*')
 lexeme = lambda p: p << whitespace
-
 lbrace = lexeme(string('{'))
 rbrace = lexeme(string('}'))
 lbrack = lexeme(string('['))
@@ -16,15 +13,14 @@ comma  = lexeme(string(','))
 true   = lexeme(string('true')).result(True)
 false  = lexeme(string('false')).result(False)
 null   = lexeme(string('null')).result(None)
-
 number = lexeme(
     regex(r'-?(0|[1-9][0-9]*)([.][0-9]+)?([eE][+-]?[0-9]+)?')
 ).map(float)
-
 string_part = regex(r'[^"\\]+')
 string_esc = string('\\') >> (
     string('\\')
     | string('/')
+    | string('"')
     | string('b').result('\b')
     | string('f').result('\f')
     | string('n').result('\n')
@@ -32,17 +28,10 @@ string_esc = string('\\') >> (
     | string('t').result('\t')
     | regex(r'u[0-9a-fA-F]{4}').map(lambda s: chr(int(s[1:], 16)))
 )
+quoted = lexeme(string('"') >> (string_part | string_esc).many().concat() << string('"'))
 
 
-@lexeme
-@generate
-def quoted():
-    yield string('"')
-    body = yield (string_part | string_esc).many()
-    yield string('"')
-    return ''.join(body)
-
-
+# Circular dependency between array and value means we use `generate` form here
 @generate
 def array():
     yield lbrack
@@ -59,14 +48,7 @@ def object_pair():
     return (key, val)
 
 
-@generate
-def json_object():
-    yield lbrace
-    members = yield object_pair.sep_by(comma)
-    yield rbrace
-    return dict(members)
-
-
+json_object = lbrace >> object_pair.sep_by(comma).map(dict) << rbrace
 value = quoted | number | json_object | array | true | false | null
 json = whitespace >> value