+++ travis_tests.sh
@@ -1,0 +1,3 @@
+#!/usr/bin/env sh
+
+coverage run --branch --source=parsy `which py.test` && coveralls