+++ docs/history.rst
@@ -16,6 +16,7 @@ History and release notes
 * Added :func:`string_from`
 * Added :data:`any_char`
 * Added :data:`decimal_digit`
+* Added :meth:`Parser.concat`
 
 0.9.0 - 2017-09-28
 ------------------